test(builder): should ignore some e2e tests that rspack does not need attention
