+++ tests/e2e/builder/cases/tools.webpack.test.ts
@@ -1,0 +1,33 @@
+import { join } from 'path';
+import { expect, test } from '@modern-js/e2e/playwright';
+import { build, getHrefByEntryName } from '../scripts/shared';
+
+const fixtures = __dirname;
+
+test('webpackChain plugin', async ({ page }) => {
+  const builder = await build({
+    cwd: join(fixtures, 'source/global-vars'),
+    entry: {
+      main: join(fixtures, 'source/global-vars/src/index.ts'),
+    },
+    runServer: true,
+    builderConfig: {
+      tools: {
+        webpackChain: (chain, { webpack }) => {
+          chain.plugin('define').use(webpack.DefinePlugin, [
+            {
+              ENABLE_TEST: JSON.stringify(true),
+            },
+          ]);
+        },
+      },
+    },
+  });
+
+  await page.goto(getHrefByEntryName('main', builder.port));
+  await expect(
+    page.evaluate(`document.getElementById('test-el').innerHTML`),
+  ).resolves.toBe('aaaaa');
+
+  builder.close();
+});
