+++ tests/e2e/builder/playwright.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from '@playwright/test';
+
+export default defineConfig({
+  // exclude webpack / rspack self-feature test cases when run rspack / webpack test
+  testIgnore:
+    process.env.PROVIDE_TYPE === 'rspack'
+      ? '**/cases/**/**.webpack.test.ts'
+      : '**/cases/**/**.rspack.test.ts',
+});
